// client/src/models/items/armours/armour.types.ts
export enum ArmourType {
  VEST = 'vest',
  VEST2 = 'vest 2',
  VEST3 = 'vest 3',
  COTTON_VEST = 'cotton vest',
  COTTON_VEST2 = 'cotton vest 2',
  COTTON_VEST3 = 'cotton vest 3',
  LEATHER_VEST = 'leather vest',
  LEATHER_VEST2 = 'leather vest 2',
  LEATHER_VEST3 = 'leather vest 3',
  ARMOUR = 'armour',
  ARMOUR2 = 'armour 2',
  ARMOUR3 = 'armour 3',
  IRON_ARMOUR = 'iron armour',
  IRON_ARMOUR2 = 'iron armour 2',
  IRON_ARMOUR3 = 'iron armour 3',
  BRONZE_ARMOUR = 'bronze armour',
  BRONZE_ARMOUR2 = 'bronze armour 2',
  BRONZE_ARMOUR3 = 'bronze armour 3',
  SILVER_ARMOUR = 'silver armour',
  SILVER_ARMOUR2 = 'silver armour 2',
  SILVER_ARMOUR3 = 'silver armour 3',
  GOLD_ARMOUR = 'gold armour',
  GOLD_ARMOUR2 = 'gold armour 2',
  GOLD_ARMOUR3 = 'gold armour 3',
  TURQUOISE_ARMOUR = 'turquoise armour',
  TURQUOISE_ARMOUR2 = 'turquoise armour 2',
  TURQUOISE_ARMOUR3 = 'turquoise armour 3',
  EMERALD_ARMOUR = 'emerald armour',
  EMERALD_ARMOUR2 = 'emerald armour 2',
  EMERALD_ARMOUR3 = 'emerald armour 3',
}